12. 포스트 작성 리팩토링
