+++ src/main/java/com/nexters/phochak/post/application/port/in/HashtagUseCase.java
@@ -2,15 +2,15 @@ package com.nexters.phochak.post.application.port.in;
 
 import com.nexters.phochak.post.adapter.out.persistence.Hashtag;
 import com.nexters.phochak.post.adapter.out.persistence.HashtagFetchDto;
-import com.nexters.phochak.post.adapter.out.persistence.PostEntity;
+import com.nexters.phochak.post.domain.Post;
 
 import java.util.List;
 import java.util.Map;
 
 public interface HashtagUseCase {
-    List<Hashtag> saveHashtags(List<String> stringHashtagList, PostEntity postEntity);
+    List<Hashtag> saveHashtags(Post post, List<String> stringHashtagList);
 
     Map<Long, HashtagFetchDto> findHashtagsOfPosts(List<Long> postIds);
 
-    void updateAll(PostEntity postEntity, List<String> stringHashtagList);
+    void updateAll(Post post, List<String> stringHashtagList);
 }
