+++ src/main/java/com/nexters/phochak/shorts/application/ShortsUseCase.java
@@ -1,11 +1,11 @@
 package com.nexters.phochak.shorts.application;
 
-import com.nexters.phochak.post.adapter.out.persistence.PostEntity;
+import com.nexters.phochak.post.domain.Post;
 import com.nexters.phochak.shorts.EncodingCallbackRequestDto;
 
 public interface ShortsUseCase {
 
-    void connectShorts(String key, PostEntity postEntity);
+    void connectShorts(Post postEntity, String key);
 
     void processPost(EncodingCallbackRequestDto encodingCallbackRequestDto);
 }
